@@ -3,6 +3,7 @@ package com.pdl.studyai_backend.service;
 import java.util.List;
 import java.util.Optional;
 
+// import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.pdl.studyai_backend.model.User;
@@ -11,25 +12,28 @@ import com.pdl.studyai_backend.repository.UserRepository;
 @Service
 public class UserService {
 
-    private final UserRepository repo;
+    private final UserRepository userRepository;
+    // private final BCryptPasswordEncoder passwordEncoder;
 
-    public UserService(UserRepository repo) {
-        this.repo = repo;
+    public UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+        // this.passwordEncoder = new BCryptPasswordEncoder();
     }
 
     public User create(User user) {
-        return repo.save(user);
+        // user.setPassword(passwordEncoder.encode(user.getPassword()));
+        return userRepository.save(user);
     }
 
     public Optional<User> findByEmail(String email) {
-        return repo.findByEmail(email);
+        return userRepository.findByEmail(email);
     }
 
     public List<User> findAll() {
-        return repo.findAll();
+        return userRepository.findAll();
     }
 
     public boolean existsByEmail(String email) {
-        return repo.existsByEmail(email);
+        return userRepository.existsByEmail(email);
     }
 }
