@@ -1,4 +1,9 @@
-// src/lib/utils.ts
-export function cn(...classes: Array<string | false | null | undefined>) {
-    return classes.filter(Boolean).join(" ");
+import { clsx, type ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
+export function generateUniqueId(prefix = ""): string {
+  return prefix + crypto.randomUUID();
 }
